consistent-type-imports rule (#2619)
related #2618
revives https://github.com/cursorless-dev/cursorless/pull/1832

vscode quick fix import and auto import appears to be doing the right
thing in regards to type only or runtime imports. Imports organize or
imports fix unfortunately do not convert a runtime import to a type
import if it's only used as a type. We probably need to rely on the
linter to do that part for us.

## Checklist

- [/] I have added
[tests](https://www.cursorless.org/docs/contributing/test-case-recorder/)
- [/] I have updated the
[docs](https://github.com/cursorless-dev/cursorless/tree/main/docs) and
[cheatsheet](https://github.com/cursorless-dev/cursorless/tree/main/cursorless-talon/src/cheatsheet)
- [/] I have not broken the cheatsheet

---------

Co-authored-by: pre-commit-ci-lite[bot] <117423508+pre-commit-ci-lite[bot]@users.noreply.github.com>