Add item scope to JS (#2675)
```js
const value = [1, 2];
```
This example is problematic. Currently we test the language specific
implementation before the text based one. This means that if your cursor
is inside the curly brackets `item` is `value = [1, 2]`. I don't think
this can be solved until we have a better `oneOf` implementation of
collection item.
Same problem with Java
https://github.com/cursorless-dev/cursorless/pull/2674

To fix that above I made a change so that the text based item can be
used even if there is a syntax tree one available if it's smaller. This
needs to merge before the java pr.

## Checklist

- [x] I have added
[tests](https://www.cursorless.org/docs/contributing/test-case-recorder/)
- [/] I have updated the
[docs](https://github.com/cursorless-dev/cursorless/tree/main/docs) and
[cheatsheet](https://github.com/cursorless-dev/cursorless/tree/main/cursorless-talon/src/cheatsheet)
- [/] I have not broken the cheatsheet