Increase performance for Tree sitter matches (#2713)
First, were checking the captures against the SCM files in production,
which was an oversight - should only be happening in debugging.

Secondarily, the existing implementation multiple map and filters. This
part of the code is very time sensitive and can run quite long for large
files since we are iterating every capture in the entire file.

This implementation aims to reduce iteration steps

Fixes https://github.com/cursorless-dev/cursorless/issues/2656

## Checklist

- [/] I have added
[tests](https://www.cursorless.org/docs/contributing/test-case-recorder/)
- [/] I have updated the
[docs](https://github.com/cursorless-dev/cursorless/tree/main/docs) and
[cheatsheet](https://github.com/cursorless-dev/cursorless/tree/main/cursorless-talon/src/cheatsheet)
- [/] I have not broken the cheatsheet