Disable Cursorless for vscode file dialogs (#2749)
File dialogs interfere with the command server keypress. Commands like
`"paste to line"` just ends up with a timeout error in the talon log and
nothing happens.

I have been running with this in my own fork of cursorless Talon for a
while to test it out and I'm happy with the result.

## Checklist

- [/] I have added
[tests](https://www.cursorless.org/docs/contributing/test-case-recorder/)
- [/] I have updated the
[docs](https://github.com/cursorless-dev/cursorless/tree/main/docs) and
[cheatsheet](https://github.com/cursorless-dev/cursorless/tree/main/cursorless-talon/src/cheatsheet)
- [/] I have not broken the cheatsheet

---------

Co-authored-by: Phil Cohen <phillip@phillip.io>