Added voice command to migrate Cursorless snippet to community format (#2747)
2 fields in the Cursorless snippets are not available in the community
format

1. scopeTypes. To my knowledge the only use case is omitting the
`function` keyword when in a javascript class
2. excludeDescendantScopeTypes: Used in conjunction with `scopeTypes` to
include the function keyword when in a function inside a class.

scopeTypes and excludeDescendantScopeTypes I'm not to sure we want to
add to community. First of all they have a single use case and needing
two separate fields just to describe this behavior fields a bit verbose.
As well community doesn't have support for sending multiple snippets.

I'm planning this as a follow up.

Reference

https://github.com/cursorless-dev/cursorless/blob/57acd6c0fcc32e5df122ce09bb26bde5f09495ca/packages/common/src/types/snippet.types.ts#L4-L24

Fixes #2355

## Checklist

- [ ] I have added
[tests](https://www.cursorless.org/docs/contributing/test-case-recorder/)
- [x] I have updated the
[docs](https://github.com/cursorless-dev/cursorless/tree/main/docs) and
[cheatsheet](https://github.com/cursorless-dev/cursorless/tree/main/cursorless-talon/src/cheatsheet)
- [/] I have not broken the cheatsheet

---------

Co-authored-by: pre-commit-ci-lite[bot] <117423508+pre-commit-ci-lite[bot]@users.noreply.github.com>
Co-authored-by: Phil Cohen <phillip@phillip.io>