Don't discard nodes with non adjacent error siblings (#2860)
Today we discard parse tree nodes if any of the ancestral nodes has a
error. That means that any of the ancestral nodes has a descendant that
is an error node. That could be something a hundred lines below that
isn't really relevant to the node we are evaluating. This updates
changes so we only look at error nodes that are adjacent to our
ascending path.